# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "easyOrder 3.0",
    "description": "Sistema de Gestão de Pedidos - Tech Challenge (3) da Pos Tech SOAT ",
    "version": "1.0.0"
  },
  "host": "",
  "paths": {
...
    "/pagamento/webhook/": {
      "post": {
        "tags": [
          "Pagamentos"
        ],
        "summary": "Webhook para Confirmação de Pagamento (Transação)",
        "description": "Captura o retorno de status de transação a partir de webhook disparado pelo Serviço de Pagamento.<br>  - Pode ser utilizado para simulação, nesse caso informe o id e status de aprovação no corpo da requisição<br><br>  [ Endpoint para integração ao gateway de pagamentos - Externo]",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "string",
              "required": [
                "id",
                "status"
              ],
              "properties": {
                "id": {
                  "type": "string",
                  "minLength": 36,
                  "maxLength": 36,
                  "format": "uuid",
                  "example": "29a81eeb-d16d-4d6c-a86c-e13597667307"
                },
                "status": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 255,
                  "example": "approved"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Solicitação Inválida"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "example": "any"
                  },
                  "status": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/pagamento/listar-transacoes/{pedidoId}": {
      "get": {
        "tags": [
          "Pagamentos"
        ],
        "summary": "Listar Transações",
        "description": "Obtém a lista de Transações associadas a um id de Pedido.<br>  - Utilizada para fins de simulação e consulta de transações pela adninistração.<br><br>  [ Endpoint para integração aos sistemas administrativo e/ou de loja ]",
        "parameters": [
          {
            "name": "pedidoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Solicitação Inválida"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  }
}